fix(hmpb): line `all-bubble-ballot` bubbles up with the grid
